refactor: 디렉터리 변경
- Servlet, Repository 디렉터리 이동
