+++ src/main/java/cafe/AppContextListener.java
@@ -3,10 +3,10 @@ package cafe;
 import cafe.questions.QuestionServlet;
 import cafe.questions.QuestionWriteServlet;
 import cafe.questions.QuestionsServlet;
-import cafe.users.UserEditServlet;
-import cafe.users.UserRegisterServlet;
-import cafe.users.UsersProfileServlet;
-import cafe.users.UsersServlet;
+import cafe.users.servlet.UserEditServlet;
+import cafe.users.servlet.UserRegisterServlet;
+import cafe.users.servlet.UsersProfileServlet;
+import cafe.users.servlet.UsersServlet;
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.ServletContextEvent;
 import jakarta.servlet.ServletContextListener;
