+++ src/main/java/cafe/users/repository/UserRepository.java
@@ -1,4 +1,6 @@
-package cafe.users;
+package cafe.users.repository;
+
+import cafe.users.User;
 
 import java.util.List;
 
@@ -8,4 +10,6 @@ public interface UserRepository {
     List<User> findAll();
 
     User findById(Long id);
+
+    void deleteAll();
 }
