+++ src/main/java/cafe/users/servlet/UserEditServlet.java
@@ -1,6 +1,8 @@
-package cafe.users;
+package cafe.users.servlet;
 
 import cafe.MappingHttpServlet;
+import cafe.users.User;
+import cafe.users.repository.UserRepository;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
